ignores and a first pass on the tinyos cc2520rpi driver
I've gotten the cc2520rpi driver to compile...but that's about it. I'm not sure all of the files I created will make it into the final version as I'm not quite sure what is handled in tinyos and what's in the driver. I assume much of what I'm doing now will have to be replaced or moved, but that is OK.

Oh, and FUCK SVN.


git-svn-id: svn+ssh://energy.eecs.umich.edu/svnroot/shed/tinyos/support/make@2763 d5a83394-2863-40c7-9603-03341d4c7fdc